requirements.txt
tweepy
pytz
tda-api
webdriver_manager